requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-types
async-timeout
attrs
blinker
Bottleneck
Brotli
cachetools
certifi
cffi
charset-normalizer
click
colorama
cryptography
frozenlist
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
google-genai # Add new SDK
googleapis-common-protos
grpcio
grpcio-status==1.71.0
httplib2
idna
Jinja2
jsonschema
jsonschema-specifications
MarkupSafe
mkl-service==2.4.0
mkl_fft
mkl_random
multidict
narwhals
numexpr
numpy
packaging
pandas
pillow
propcache
proto-plus
protobuf
pyarrow
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
pyOpenSSL
pyparsing
PySocks
python-dateutil
python-dotenv==1.1.0
pytz
pyu2f
PyYAML
referencing
requests
rpds-py
rsa
setuptools==72.1.0
six
smmap
streamlit
tenacity
toml
tornado
tqdm
typing-inspection==0.4.1
typing_extensions
tzdata
uritemplate
urllib3
watchdog
wheel==0.45.1
yarl
